Move InferPulumiValue to crosstestsimpl
InferPulumiValue was made public in a previous commit, but it shouldn't really be exposed
to end users of cross-tests, so this commit moved it to it's own package.
