+++ pkg/internal/tests/cross-tests/diff_check.go
@@ -29,6 +29,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	crosstestsimpl "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/internal/tests/cross-tests/impl"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/pulcheck"
 )
 
@@ -85,12 +86,12 @@ func runDiffCheck(t T, tc diffTestCase) diffResult {
 		tfResourceName:      defRtype,
 	}
 
-	yamlProgram := pd.generateYAML(t, InferPulumiValue(t,
+	yamlProgram := pd.generateYAML(t, crosstestsimpl.InferPulumiValue(t,
 		bridgedProvider.P.ResourcesMap().Get(defRtype).Schema(), nil, tfConfig1))
 	pt := pulcheck.PulCheck(t, bridgedProvider, string(yamlProgram))
 	pt.Up(t)
 
-	yamlProgram = pd.generateYAML(t, InferPulumiValue(t,
+	yamlProgram = pd.generateYAML(t, crosstestsimpl.InferPulumiValue(t,
 		bridgedProvider.P.ResourcesMap().Get(defRtype).Schema(), nil, tfConfig2))
 	err := os.WriteFile(filepath.Join(pt.CurrentStack().Workspace().WorkDir(), "Pulumi.yaml"), yamlProgram, 0o600)
 	require.NoErrorf(t, err, "writing Pulumi.yaml")
