+++ pkg/internal/tests/cross-tests/puwrite_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v3"
 
+	crosstestsimpl "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/internal/tests/cross-tests/impl"
 	shimv2 "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim/sdk-v2"
 )
 
@@ -124,7 +125,7 @@ runtime: yaml
 			))
 			schema := shimProvider.ResourcesMap().Get(rtype).Schema()
 			out, err := generateYaml(t, rtoken,
-				InferPulumiValue(t, schema, nil, coalesceInputs(t, tc.schema, tc.tfConfig)))
+				crosstestsimpl.InferPulumiValue(t, schema, nil, coalesceInputs(t, tc.schema, tc.tfConfig)))
 			require.NoError(t, err)
 			b, err := yaml.Marshal(out)
 			require.NoError(t, err)
