+++ pkg/internal/tests/cross-tests/upgrade_state_check.go
@@ -29,6 +29,7 @@ import (
 	"github.com/zclconf/go-cty/cty"
 	"gotest.tools/v3/assert"
 
+	crosstestsimpl "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/internal/tests/cross-tests/impl"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tests/pulcheck"
 )
 
@@ -87,14 +88,14 @@ func runPulumiUpgrade(t T, res1, res2 *schema.Resource, config1, config2 cty.Val
 		tfResourceName:      defRtype,
 	}
 
-	yamlProgram := pd.generateYAML(t, InferPulumiValue(t,
+	yamlProgram := pd.generateYAML(t, crosstestsimpl.InferPulumiValue(t,
 		prov1.P.ResourcesMap().Get(pd.tfResourceName).Schema(), nil, config1))
 	pt := pulcheck.PulCheck(t, prov1, string(yamlProgram))
 	pt.Up(t)
 	stack := pt.ExportStack(t)
 	schemaVersion1 := getVersionInState(t, stack)
 
-	yamlProgram = pd.generateYAML(t, InferPulumiValue(t,
+	yamlProgram = pd.generateYAML(t, crosstestsimpl.InferPulumiValue(t,
 		prov1.P.ResourcesMap().Get(pd.tfResourceName).Schema(), nil, config2))
 	p := filepath.Join(pt.CurrentStack().Workspace().WorkDir(), "Pulumi.yaml")
 	err := os.WriteFile(p, yamlProgram, 0o600)
